requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
pytest==7.1.3
pytest-django==4.5.2
psycopg2-binary==2.9.3
python-dotenv==0.21.0
